+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `GeolocationInput` optional `autocompleteOptions` prop.
+- `loadingRules` to `AddressRulesContext` to enhance UX possibilities.
 
 ## [4.8.0] - 2021-08-03